renamed HTMLExportDialog to DocumentationExportDialog
